refactor: moved to utils/clients/constants & applied new directory
