+++ components/trending/module/TrendingCategoriesList.tsx
@@ -6,7 +6,7 @@ import {
   QUERY_STALE_TIME,
   RECENT_FILTER,
   RECENT_FILTER_ARR,
-} from "@/libs/utils/client/commonValues";
+} from "@/libs/utils/client/contants/commonValues";
 import { useQueries } from "@tanstack/react-query";
 import {
   getFilteredCategoriesPlaylists,
