+++ components/trending/module/TrendingMainList.tsx
@@ -7,7 +7,7 @@ import {
   QUERY_STALE_TIME,
   RECENT_FILTER,
   RECENT_FILTER_ARR,
-} from "@/libs/utils/client/commonValues";
+} from "@/libs/utils/client/contants/commonValues";
 import { ListBulletIcon, Squares2X2Icon } from "@heroicons/react/24/solid";
 import ReactQueryErrorBoundary from "@/components/common/react-query-provider/ReactQueryErrorBoundary";
 import PlaylistListContainer from "@/components/common/playlist/column-list/PlaylistListContainer";
