+++ libs/utils/client/contants/commonValues.ts
@@ -1,6 +1,5 @@
 import {
   PlaylistMoodType,
-  PlaylistType,
   SONG_URL_TYPE,
 } from "@/libs/types/song&playlistType";
 
@@ -10,36 +9,6 @@ export const breakpoints = {
   laptop: 1024,
   desktop: 1280,
   desktopL: 1440,
-};
-
-interface IDefaultPlaylist extends Omit<PlaylistType, "author" | "updatedAt"> {
-  author: {
-    id: string;
-    profilePic: string;
-    nickname: string;
-  };
-}
-
-export const playlistDefault: IDefaultPlaylist = {
-  title: "",
-  description: "",
-  coverImage: "",
-  createdAt: "",
-  author: {
-    id: "",
-    profilePic: "",
-    nickname: "",
-  },
-  authorId: "",
-  songs: [],
-  id: "",
-  comments: [],
-  likedBy: [],
-  playedCount: 0,
-  mood: {
-    name: "relaxed",
-  },
-  category: [],
 };
 
 export const commonMoods: PlaylistMoodType[] = [
